% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_and_preprocess_data}
\alias{import_and_preprocess_data}
\title{Importa e pré-processa dados bibliográficos}
\usage{
import_and_preprocess_data(scopus_file_path, wos_file_path)
}
\arguments{
\item{scopus_file_path}{Caminho para o arquivo .bib do Scopus.}
\item{wos_file_path}{Caminho para o arquivo .txt ou .bib do Web of Science.}
}
\value{Um data.frame mesclado e deduplicado com os dados bibliográficos.}
\description{
Esta função importa dados de Scopus e Web of Science, mescla-os e remove duplicatas.
}
\examples{
# Exemplo de uso (requer arquivos de exemplo)
# M <- import_and_preprocess_data("path/to/scopus.bib", "path/to/wos.txt")
}

