% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{perform_meta_analysis}
\alias{perform_meta_analysis}
\title{Realiza meta-análise}
\usage{
perform_meta_analysis(data_extracted, model_type)
}
\arguments{
\item{data_extracted}{Um data.frame com os dados extraídos para meta-análise.}
\item{model_type}{Tipo de modelo de efeito (e.g., "Efeitos Fixos", "Efeitos Aleatórios").}
}
\value{Um objeto meta do pacote meta com os resultados da meta-análise.}
\description{
Esta função executa uma meta-análise com base nos dados de efeito fornecidos.
}
\examples{
# Exemplo de uso (requer dados extraídos)
# meta_results <- perform_meta_analysis(my_extracted_data, "Efeitos Aleatórios")
}

