% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Lança o aplicativo Shiny do FAET}
\usage{
run_app()
}
\description{
Esta função inicia a interface gráfica do usuário do pacote FAET.
}
\examples{
if (interactive()) {
  FAET::run_app()
}
}

