% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling.R
\name{perform_topic_modeling}
\alias{perform_topic_modeling}
\title{Realiza modelagem de tópicos (LDA)}
\usage{
perform_topic_modeling(dfm, num_topics)
}
\arguments{
\item{dfm}{Um objeto dfm (document-feature matrix) da quanteda.}
\item{num_topics}{Número de tópicos (k) para o modelo LDA.}
}
\value{Um objeto LDA do pacote topicmodels.}
\description{
Esta função aplica o algoritmo Latent Dirichlet Allocation (LDA) para identificar tópicos latentes em um corpus textual.
}
\examples{
# Exemplo de uso (requer um dfm)
# lda_model <- perform_topic_modeling(corpus_dfm, 5)
}

